requirements.txt
plyer
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pillow
pync
